ugh... stupid keyboard issue!!!!
Finally a fix for the keyboard issue. It seems the fragments don't play nice when there is an EditText somewhere... it ends up taking focus when it shouldn't, and the keyboard pops up automatically.
Lots of StackedOverflow suggestions around forcing focus one way or the other failed. In the end, I had to force close the keyboard...ugh.